Keep functions for handling files/directories in seperate file
